+++ Cargo.toml
@@ -211,6 +211,7 @@ serde_ini            = "0.2.0"
 serde_json           = "1.0.139"
 similar              = "2.7.0"
 smallvec             = { version = "1.13.2", features = ["union", "const_new", "serde"] }
+static_assertions    = "1.1"
 syn                  = "1.0.109"
 termcolor            = "1.4.1"
 terminal_size        = "0.4.1"
